+++ src/pages/Admin/AdminPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SideBar from '../components/SideBar';
+import SideBar from '../../components/SideBar';
 
 export default function AdminPage() {
   return (
